Add visual differences and common vim changes
Filling in more differences from vim and how to restore vim behavior.

Many users make the same changes to vim. Spacemacs makes some of these
changes already, but not all of them. We could convert each item in this
section to a bug to make it the default spacemacs behavior.